debug

+++ Homework/Assignment3/Assignment3/rasterizer.cpp
@@ -292,7 +292,7 @@ void rst::rasterizer::rasterize_triangle(const Triangle& t, const std::array<Eig
                  fragment_shader_payload payload(interpolated_color, interpolated_normal.normalized(), interpolated_texcoords, texture ? &*texture : nullptr);
                  payload.view_pos = interpolated_shadingcoords;
                  auto pixel_color = fragment_shader(payload);
-                 Eigen::Vector2d point = { x,y };
+                 Eigen::Vector2i point = { x,y };
                  set_pixel(point, pixel_color);
              }
         }
